SCRUM 18
add amount to user when tap is done

+++ BeerTap/Repositories/UserRepository.cs
@@ -150,7 +150,7 @@
             }
         }
 
-        public async Task UpdateUserScoreAsync(string userId, int newScore)
+        public async Task UpdateUserScoreAsync(string userId, float newScore)
         {
             using (var sqlConnection = new SqlConnection(sqlConnectionString))
             {
